CCK: refactoring of test code
