+++ cck/shared_examples.rb
@@ -2,6 +2,9 @@ require 'json'
 require 'rspec'
 require 'cucumber/messages'
 
+require_relative 'lib/messages_comparator'
+require_relative 'lib/keys_checker'
+
 def message_type(message)
   message.to_hash.keys.first
 end
@@ -23,88 +26,6 @@ def debug_lists(expected, obtained)
   end
 end
 
-class MessagesComparator
-  attr_reader :errors
-
-  def initialize(found, expected)
-    @errors = []
-    @compared = []
-
-    compare(found, expected)
-  end
-
-  def debug
-    puts 'Compared the following type of message:'
-    puts @compared.uniq.map { |m| " - #{m}" }.join("\n")
-    puts ''
-    puts errors.uniq.join("\n")
-  end
-
-  private
-
-  def compare(found, expected)
-    found_by_type = messages_by_type(found)
-    expected_by_type = messages_by_type(expected)
-
-    found_by_type.keys.each do |type|
-      compare_list(found_by_type[type], expected_by_type[type])
-    end
-  end
-
-  def messages_by_type(messages)
-    by_type = Hash.new { |h, k| h[k] = [] }
-    messages.each do |msg|
-      by_type[message_type(msg)] << remove_envelope(msg)
-    end
-    by_type
-  end
-
-  def remove_envelope(message)
-    message[message_type(message)]
-  end
-
-  def compare_list(found, expected)
-    found.each_with_index do |message, index|
-      compare_message(message, expected[index])
-    end
-  end
-
-  def compare_message(found, expected)
-    return unless found.is_a?(Protobuf::Message)
-    return if found.is_a?(Cucumber::Messages::GherkinDocument)
-    return if found.is_a?(Cucumber::Messages::Pickle)
-
-    @compared << found.class.name
-    same_keys?(found, expected)
-    compare_sub_messages(found, expected)
-  end
-
-  def compare_sub_messages(found, expected)
-    return unless expected.respond_to? :to_hash
-    expected.to_hash.keys.each do |key|
-      value = expected.send(key)
-      if value.is_a?(Protobuf::Message)
-        compare_message(found.send(key), value)
-      elsif value.is_a?(Array)
-        compare_list(found.send(key), value)
-      end
-    end
-  end
-
-  def same_keys?(found, expected)
-    found_keys = found.to_hash.keys
-    expected_keys = expected.to_hash.keys
-
-    return if found_keys.sort == expected_keys.sort
-
-    missing_keys = expected_keys - found_keys
-    extra_keys = found_keys - expected_keys
-
-    errors << "Found extra keys in message #{found.class.name}: #{extra_keys}" unless extra_keys.empty?
-    errors << "Missing keys in message #{found.class.name}: #{missing_keys}" unless missing_keys.empty?
-  end
-end
-
 RSpec.shared_examples 'equivalent messages' do
   # Note: to use those examples, you need to define:
   # let(:original) { 'path to .ndjson file in CCK' }
@@ -117,12 +38,12 @@ RSpec.shared_examples 'equivalent messages' do
   let(:generated_messages_types) { parsed_generated.map { |msg| message_type(msg) } }
 
   it 'produces the same kind of messages' do
-    debug_lists(original_messages_types, generated_messages_types)
+    # debug_lists(original_messages_types, generated_messages_types)
     expect(generated_messages_types).to contain_exactly(*original_messages_types)
   end
 
-  it 'produces messages with the same content' do
-    comparator = MessagesComparator.new(parsed_generated, parsed_original)
+  it 'produces messages with the same keys' do
+    comparator = CCK::MessagesComparator.new(CCK::KeysChecker, parsed_generated, parsed_original)
     comparator.debug if ENV['VERBOSE']
 
     expect(comparator.errors).to be_empty
